app/requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
openai==0.27.6
python-dotenv==1.0.0
langchain==0.0.191
Chroma==0.2.0
chromadb==0.3.26